Methods list

SPECTrecon.SPECTrecon

SPECTrecon.Rotate1D

SPECTrecon.Rotate2D

SPECTrecon.PlanPSF

SPECTrecon.PlanRotate

SPECTrecon.RotateMode

Base.show

Base.show

Base.sizeof

SPECTrecon.backproject

SPECTrecon.backproject

SPECTrecon.backproject!

SPECTrecon.backproject!

SPECTrecon.backproject!

SPECTrecon.copy3dj!

SPECTrecon.fft_conv

SPECTrecon.fft_conv!

SPECTrecon.fft_conv_adj

SPECTrecon.fft_conv_adj!

SPECTrecon.fftshift2!

SPECTrecon.imfilterz!

SPECTrecon.imrotate

SPECTrecon.imrotate!

SPECTrecon.imrotate!

SPECTrecon.imrotate!

SPECTrecon.imrotate_adj

SPECTrecon.imrotate_adj!

SPECTrecon.imrotate_adj!

SPECTrecon.imrotate_adj!

SPECTrecon.linearinterp!

SPECTrecon.mul3dj!

SPECTrecon.pad2sizezero!

SPECTrecon.pad_it!

SPECTrecon.padrepl!

SPECTrecon.padzero!

SPECTrecon.plan_psf

SPECTrecon.plan_rotate

SPECTrecon.plus1di!

SPECTrecon.plus1dj!

SPECTrecon.plus2di!

SPECTrecon.plus2dj!

SPECTrecon.plus3di!

SPECTrecon.plus3dj!

SPECTrecon.plus3dk!

SPECTrecon.project

SPECTrecon.project

SPECTrecon.project!

SPECTrecon.project!

SPECTrecon.project!

SPECTrecon.rot180!

SPECTrecon.rot_f90!

SPECTrecon.rot_f90_adj!

SPECTrecon.rotate_x!

SPECTrecon.rotate_x_adj!

SPECTrecon.rotate_y!

SPECTrecon.rotate_y_adj!

SPECTrecon.rotl90!

SPECTrecon.rotr90!

SPECTrecon.scale3dj!

Methods usage

SPECTrecon.SPECTrecon — Module

SPECTrecon

System matrix (forward and back-projector) for SPECT image reconstruction.

SPECTrecon.Rotate1D — Constant

Rotate1D = RotateMode{:one}()

Rotate using 3-pass 1D interpolation

SPECTrecon.Rotate2D — Constant

Rotate2D = RotateMode{:two}()

Rotate using 1-pass 2D bilinear interpolation

SPECTrecon.PlanPSF — Type

PlanPSF(nx::Int, nz::Int, nx_psf::Int; T::DataType)

Make struct for storing work arrays and factors for 2D convolution for one thread Currently PSF must be square, i.e., nx_psf = nz_psf

T datatype of work arrays

nx must be Int

nz must be Int

nx_psf must be Int

padsize{padup, paddown, padleft, padright} Tuple of padsize

workmat [nx+padup+paddown, nz+padleft+padright] 2D padded image for FFT convolution

workvecx [nx+padup+paddown,]: 1D work vector

workvecz [nz+padleft+padright,]: 1D work vector

img_compl [nx+padup+paddown, nz+padleft+padright]: 2D [complex] padded image for fft

ker_compl [nx+padup+paddown, nz+padleft+padright]: 2D [complex] padded image for fft

fft_plan plan for doing fft, see plan_fft!

ifft_plan plan for doing ifft, see plan_ifft!

SPECTrecon.PlanRotate — Type

PlanRotate{T, R}

Struct for storing work arrays and factors for 2D square image rotation for one thread

T datatype of work arrays (default Float32)

R::RotateMode

nx::Int image size

padsize::Int : pad each side of image by this much

workmat1 [nx + 2 * padsize, nx + 2 * padsize] padded work matrix

workmat2 [nx + 2 * padsize, nx + 2 * padsize] padded work matrix

workvec [nx + 2 * padsize, ] padded work vector

interp::SparseInterpolator{T, 2, 1}

SPECTrecon.RotateMode — Type

RotateMode

Type to control image rotation method.

Rotate1D : 3-pass 1D interpolation

Rotate2D : 1-pass 2D bilinear interpolation

Base.show — Method

show(io::IO, mime::MIME"text/plain", vp::Vector{<:PlanRotate})

Base.show — Method

show(io::IO, ::MIME"text/plain", plan::PlanRotate)

Base.sizeof — Method

sizeof(::PlanRotate)

Show size in bytes of PlanRotate object.

SPECTrecon.backproject! — Method

backproject!(image, view, plan, viewidx)

Backproject a single view.

SPECTrecon.backproject! — Method

backproject!(image, views, plan ; index)

Backproject multiple views into image. Array image is not initialized to zero; caller must do that.

SPECTrecon.backproject! — Method

backproject!(image, view, plan, thid, viewidx)

Backproject a single view.

SPECTrecon.backproject — Method

image = backproject(views, mumap, psfs, dy; interpmeth, kwargs...)

SPECT backproject views using attenuation map mumap and PSF array psfs for pixel size dy. This method initializes the plan as a convenience. Most users should use backproject! instead after initializing those, for better efficiency.

SPECTrecon.backproject — Method

image = backproject(views, plan ; kwargs...)

SPECT backproject views; this allocates the returned 3D array.

SPECTrecon.copy3dj! — Method

copy3dj!(mat2d, mat3d, j)

Non-allocating mat2d .= mat3d[:,j,:]

SPECTrecon.fft_conv! — Method

fft_conv!(output, img, ker, plan)

Convolve img with ker using FFT, and store the result in output

SPECTrecon.fft_conv — Method

fft_conv(img, ker)

Convolve img with ker using FFT

SPECTrecon.fft_conv_adj! — Method

fft_conv_adj!(output, img, ker, plan)

Adjoint of convolving img with ker using FFT, and store the result in output

SPECTrecon.fft_conv_adj — Method

fft_conv_adj(img, ker)

Adjoint of convolving img with ker using FFT

SPECTrecon.fftshift2! — Method

fftshift2!(dst, src)

Same as fftshift in 2d, but non-allocating

SPECTrecon.imfilterz! — Method

imfilterz!(plan)

FFT-based convolution between plan.img_compl and kernel plan.ker_compl (not centered) putting result in plan.workmat.

SPECTrecon.imrotate! — Method

imrotate!(output, image3, θ, plans)

In-place version of rotating a 3D image3 by θ in counter-clockwise direction (opposite to imrotate in Julia)

SPECTrecon.imrotate! — Method

imrotate!(output, img, θ, plan)

Rotate an image by angle θ (must be within 0 to 2π) in counter-clockwise direction (opposite to imrotate in Julia) using 3-pass 1d linear interpolations.

SPECTrecon.imrotate! — Method

imrotate!(output, img, θ, plan)

Rotate an image by angle θ (must be within 0 to 2π) in counter-clockwise direction (opposite to imrotate in Julia) using 2d bilinear interpolation.

SPECTrecon.imrotate — Method

imrotate(img, θ; method::Symbol=:two)

Rotate an image by angle θ (must be within 0 to 2π) in counter-clockwise direction (opposite to imrotate in Julia) using either 2d linear interpolation (for :two) or 3-pass 1D interpolation (for :one)

SPECTrecon.imrotate_adj! — Method

imrotate_adj!(output, image3, θ, plan)

In-place version of the adjoint of rotating a 3D image3 by θ in counter-clockwise direction (opposite to imrotate in Julia)

SPECTrecon.imrotate_adj! — Method

imrotate_adj!(output, img, θ, plan)

The adjoint of rotating an image by angle θ (must be within 0 to 2π) in counter-clockwise direction (opposite to imrotate in Julia) using 3-pass 1d linear interpolations.

SPECTrecon.imrotate_adj! — Method

imrotate_adj!(output, img, θ, plan)

The adjoint of rotating an image by angle θ (must be within 0 to 2π) in counter-clockwise direction (opposite to imrotate in Julia) using 2d bilinear interpolations.

SPECTrecon.imrotate_adj — Method

imrotate_adj(img, θ; method::Symbol=:two)

The adjoint of rotating an image by angle θ (must be within 0 to 2π) in counter-clockwise direction (opposite to imrotate in Julia) using either 2d linear interpolations or 3-pass 1D interpolation.

SPECTrecon.linearinterp! — Method

linearinterp!(A, x)

Assign key values in SparseInterpolator (linear) A that are calculated from x. x must be a constant vector

SPECTrecon.mul3dj! — Method

mul3dj!(mat3d, mat2d, j)

Non-allocating mat3d[:,j,:] *= mat2d

SPECTrecon.pad2sizezero! — Method

pad2sizezero!(output, img, padsize)

Non-allocating version of padding: `output[paddims[1]+1 : paddims[1]+dims[1], paddims[2]+1 : paddims[2]+dims[2]]) .= img

SPECTrecon.pad_it! — Method

pad_it!(X, padsize)

Zero-pad X to padsize

SPECTrecon.padrepl! — Method

padrepl!(output, img, padsize)

Pad with replication from img into output

SPECTrecon.padzero! — Method

padzero!(output, img, pad_x, pad_y)

Mutating version of padding a 2D image by filling zeros. Output has size (size(img, 1) + padsize[1] + padsize[2], size(img, 2) + padsize[3] + padsize[4]).

SPECTrecon.plan_psf — Method

plan_psf(nx::Int, nz::Int, nx_psf::Int; nthread::Int, T::DataType)

Make Vector of structs for storing work arrays and factors for 2D convolution Input

nx::Int

nz::Int

nx_psf::Int

Option

T : datatype of work arrays, defaults to Float32

nthread::Int # of threads, defaults to Threads.nthreads()

SPECTrecon.plan_rotate — Method

plan_rotate(nx::Int; nthread::Int, T::DataType, method::Symbol)

Make Vector of PlanRotate structs for storing work arrays and factors for 2D square image rotation.

Input

nx::Int must equal to ny

Option

T : datatype of work arrays, defaults to Float32

method::Symbol : default is :two for 2D interpolation; use :one for 3-pass rotation with 1D interpolation

nthread::Int # of threads, defaults to Threads.nthreads() warning: must use that default currently!

SPECTrecon.plus1di! — Method

plus1di!(mat2d, mat1d)

Non-allocating mat2d[i, :] += mat1d

SPECTrecon.plus1dj! — Method

plus1dj!(mat2d, mat1d)

Non-allocating mat2d[:, j] += mat1d

SPECTrecon.plus2di! — Method

plus2di!(mat1d, mat2d, i)

Non-allocating mat1d += mat2d[i,:]

SPECTrecon.plus2dj! — Method

plus2dj!(mat1d, mat2d, j)

Non-allocating mat1d += mat2d[:,j]

SPECTrecon.plus3di! — Method

plus3di!(mat2d, mat3d, i)

Non-allocating mat2d += mat3d[i,:,:]

SPECTrecon.plus3dj! — Method

plus3dj!(mat2d, mat3d, j)

Non-allocating mat2d += mat3d[:,j,:]

SPECTrecon.plus3dk! — Method

plus3dk!(mat2d, mat3d, k)

Non-allocating mat2d += mat3d[:,:,k]

SPECTrecon.project! — Method

project!(views, image, plan; index)

Project image into multiple views with indexes index (defaults to 1:nview). The 3D views array must be pre-allocated, but need not be initialized.

SPECTrecon.project! — Method

project!(view, plan, image, viewidx)

SPECT projection of image into a single view with index viewidx. The view must be pre-allocated but need not be initialized to zero.

SPECTrecon.project! — Method

project!(view, plan, image, thid, viewidx)

SPECT projection of image into a single view with index viewidx. The view must be pre-allocated but need not be initialized to zero.

SPECTrecon.project — Method

views = project(image, mumap, psfs, dy; interpmeth, kwargs...)

Convenience method for SPECT forward projector that does all allocation including initializing plan.

In

image : 3D array [nx,ny,nz]

mumap : [nx,ny,nz] 3D attenuation map, possibly zeros()

psfs : 4D PSF array

dy::RealU : pixel size

Option

interpmeth : :one or :two

SPECTrecon.project — Method

views = project(image, plan ; kwargs...)

Convenience method for SPECT forward projector that allocates and returns views.

SPECTrecon.rot180! — Method

rot180!(B::AbstractMatrix, A::AbstractMatrix)

In place version of rot180, returning rotation of A in B.

SPECTrecon.rot_f90! — Method

rot_f90!(output, img, m)

In-place version of rotating an image by 90/180/270 degrees

SPECTrecon.rot_f90_adj! — Method

rot_f90_adj!(output, img, m)
The adjoint of rotating an image by 90/180/270 degrees

SPECTrecon.rotate_x! — Method

rotate_x!(output, img, tan_θ, workvec, interp)

SPECTrecon.rotate_x_adj! — Method

rotate_x_adj!(output, img, tan_θ, workvec, interp)

SPECTrecon.rotate_y! — Method

rotate_y!(output, img, sin_θ, workvec, interp)

SPECTrecon.rotate_y_adj! — Method

rotate_y_adj!(output, img, sin_θ, workvec, interp)

SPECTrecon.rotl90! — Method

rotl90!(B::AbstractMatrix, A::AbstractMatrix)

In place version of rotl90, returning rotation of A in B.

SPECTrecon.rotr90! — Method

rotr90!(B::AbstractMatrix, A::AbstractMatrix)

In place version of rotr90, returning rotation of A in B.

SPECTrecon.scale3dj! — Method

scale3dj!(mat2d, mat3d, j, s)

Non-allocating mat2d = s * mat3d[:,j,:]