Methods list

SPECTrecon.SPECTrecon

SPECTrecon.SPECTplan

SPECTrecon.SPECTplan

SPECTrecon.Workarray

SPECTrecon.backproject

SPECTrecon.backproject

SPECTrecon.backproject!

SPECTrecon.backproject!

SPECTrecon.copy3dj!

SPECTrecon.fft_conv!

SPECTrecon.fft_conv_adj!

SPECTrecon.fftshift2!

SPECTrecon.imfilter3!

SPECTrecon.imfilter3_adj!

SPECTrecon.imrotate3!

SPECTrecon.imrotate3!

SPECTrecon.imrotate3_adj!

SPECTrecon.imrotate3_adj!

SPECTrecon.linearinterp!

SPECTrecon.mul3dj!

SPECTrecon.pad2sizezero!

SPECTrecon.pad_it!

SPECTrecon.padrepl!

SPECTrecon.padzero!

SPECTrecon.plus1di!

SPECTrecon.plus1dj!

SPECTrecon.plus2di!

SPECTrecon.plus2dj!

SPECTrecon.plus3di!

SPECTrecon.plus3dj!

SPECTrecon.plus3dk!

SPECTrecon.project

SPECTrecon.project

SPECTrecon.project!

SPECTrecon.project!

SPECTrecon.rot180!

SPECTrecon.rot_f90!

SPECTrecon.rot_f90_adj!

SPECTrecon.rotate_x!

SPECTrecon.rotate_x_adj!

SPECTrecon.rotate_y!

SPECTrecon.rotate_y_adj!

SPECTrecon.rotl90!

SPECTrecon.rotr90!

SPECTrecon.scale3dj!

Methods usage

SPECTrecon.SPECTrecon — Module

SPECTrecon

System matrix (forward and back-projector) for SPECT image reconstruction.

SPECTrecon.SPECTplan — Type

SPECTplan

Struct for storing key factors for a SPECT system model

T datatype of work arrays

imgr [nx, ny, nz] 3D rotated version of image

add_img [nx, ny, nz] 3D image for adding views and backprojection

mumap [nx,ny,nz] attenuation map, must be 3D, possibly zeros()

mumapr [nx, ny, nz] 3D rotated mumap

psfs [nx_psf,nz_psf,ny,nview] point spread function, must be 4D, with nx_psf and nz_psf odd, and symmetric for each slice

nview number of views, must be integer

viewangle set of view angles, must be from 0 to 2π

interpidx interpolation method, 1 means 1d, 2 means 2d

dy voxel size in y direction (dx is the same value)

imgsize{nx, ny, nz} number of voxels in {x,y,z} direction of the image, must be integer

psfsize{nx_psf, nz_psf} number of voxels in {x, z} direction of the psf, must be integer

pad_fft{padu_fft, pad_fft, padl_fft, padr_fft} pixels padded for {left,right,up,down} direction for convolution with psfs, must be integer

pad_rot{padu_rot, padd_rot, padl_rot, padr_rot} padded pixels for {left,right,up,down} direction for image rotation

ncore number of CPU cores used to process data, must be integer

Currently code assumes the following:

each of the nview projection views is [nx,nz]

nx = ny

uniform angular sampling

psf is symmetric

multiprocessing using # of cores specified by Threads.nthreads()

SPECTrecon.SPECTplan — Method

SPECTplan(mumap, psfs, dy; viewangle, interpidx, T)

Constructor for SPECTplan

In

mumap::AbstractArray{<:RealU, 3} 3D attenuation map

psfs::AbstractArray{<:RealU, 4} 4D PSF array

dy::RealU pixel size

Option

viewangle::AbstractVector{<:RealU} default 0 to almost 2π

interpidx::Int = 2 1 is for 3-pass 1D interpolator; 2 is for 2D interpolator

T::DataType = promote_type(eltype(mumap), Float32)

SPECTrecon.Workarray — Type

Workarray

Struct for storing keys of the work array for a single thread add tmp vectors to avoid allocating in rotatex and rotatey For fft convolution:

workmat_fft [nx+padu_fft+padd_fft, nz+padl_fft+padr_fft]: 2D padded image for imfilter3

workvec_fft_1 [nz+padl_fft+padr_fft,]: 1D work vector

workvec_fft_2 [nx+padu_fft+padd_fft,]: 1D work vector

img_compl [nx+padu_fft+padd_fft, nz+padl_fft+padr_fft]: 2D [complex] padded image for fft

ker_compl [nx+padu_fft+padd_fft, nz+padl_fft+padr_fft]: 2D [complex] padded image for fft

fft_plan plan for doing fft, see plan_fft!

ifft_plan plan for doing ifft, see plan_ifft!

For image rotation:

workmat_rot_1 [nx+padu_rot+padd_rot, ny+padl_rot+padr_rot]: 2D padded image for image rotation

workmat_rot_2 [nx+padu_rot+padd_rot, ny+padl_rot+padr_rot]: 2D padded image for image rotation

workvec_rot_x [nx+padu_rot+padd_rot,]: 1D work vector for image rotation

workvec_rot_y [ny+padl_rot+padr_rot,]: 1D work vector for image rotation

interp_x sparse interpolator for rotating in x direction

interp_y sparse interpolator for rotating in y direction

For attenuation:

exp_mumapr [nx, nz] 2D exponential rotated mumap

SPECTrecon.backproject! — Method

backproject!(image, views, plan, workarray; index)

Backproject multiple views into image. Array image is not initialized to zero; caller must do that.

SPECTrecon.backproject! — Method

backproject!(image, view, plan, workarray, viewidx)

Backproject a single view.

SPECTrecon.backproject — Method

image = backproject(views, mumap, psfs, dy; interpidx, kwargs...)

SPECT backproject views using attenuation map mumap and PSF array psfs for pixel size dy. This method initializes the plan and workarray as a convenience. Most users should use backproject! instead after initializing those, for better efficiency.

SPECTrecon.backproject — Method

image = backproject(plan, workarray, views ; kwargs...)

SPECT backproject views; this allocates the returned 3D array.

SPECTrecon.copy3dj! — Method

copy3dj!(mat2d, mat3d, j)

Non-allocating mat2d .= mat3d[:,j,:]

SPECTrecon.fft_conv! — Method

fft_conv!(output, workmat, img, ker, fftpadsize, img_compl, ker_compl, fft_plan, ifft_plan)

Convolve img with ker using FFT, and store the result in output

SPECTrecon.fft_conv_adj! — Method

fft_conv_adj!(output, workmat, workvec1, workvec2, img, ker,
              fftpadsize, img_compl, ker_compl, fft_plan, ifft_plan)

SPECTrecon.fftshift2! — Method

fftshift2!(dst, src)

Same as fftshift in 2d, but non-allocating

SPECTrecon.imfilter3! — Method

imfilter3!(output, img_compl, ker, ker_compl, fft_plan, ifft_plan)

FFT-based convolution between img_compl and kernel ker (not centered) putting result in output.

SPECTrecon.imfilter3_adj! — Method

imfilter3_adj!(output, img_compl, kerev, ker_compl, fft_plan, ifft_plan)

Apply FFT convolution between img_compl and REVERSED kernel (not centered), assuming the kernel is already be in reversed order.

SPECTrecon.imrotate3! — Method

imrotate3!(output, workmat1, workmat2, img, θ, interp_x, interp_y, workvec_x, workvec_y)

Rotate an image by angle θ (must be within 0 to 2π) in clockwise direction using a series of 1d linear interpolations.

SPECTrecon.imrotate3! — Method

imrotate3!(output, workmat1, workmat2, img, θ)

Rotate an image by angle θ in clockwise direction using 2d linear interpolation Source code is here: https://github.com/emmt/LinearInterpolators.jl

SPECTrecon.imrotate3_adj! — Method

imrotate3_adj!(output, workmat1, workmat2, img, θ, interp_x, interp_y, workvec_x, workvec_y)

The adjoint of rotating an image by angle θ (must be ranging from 0 to 2π) in clockwise direction using a series of 1d linear interpolation

SPECTrecon.imrotate3_adj! — Method

imrotate3_adj!(output, workmat1, workmat2, img, θ)

The adjoint of rotating an image by angle θ in clockwise direction using 2d linear interpolation Source code is here: https://github.com/emmt/LinearInterpolators.jl

SPECTrecon.linearinterp! — Method

linearinterp!(A, x)

Assign key values in SparseInterpolator (linear) A that are calculated from x. x must be a constant vector

SPECTrecon.mul3dj! — Method

mul3dj!(mat3d, mat2d, j)

Non-allocating mat3d[:,j,:] *= mat2d

SPECTrecon.pad2sizezero! — Method

pad2sizezero!(output, img, padsize)

Non-allocating version of padding: `output[paddims[1]+1 : paddims[1]+dims[1], paddims[2]+1 : paddims[2]+dims[2]]) .= img

SPECTrecon.pad_it! — Method

pad_it!(X, padsize)

Zero-pad X to padsize

SPECTrecon.padrepl! — Method

padrepl!(output, img, padsize)

Pad with replication from img into output

SPECTrecon.padzero! — Method

padzero!(output, img, pad_x, pad_y)

Mutating version of padding a 2D image by filling zeros. Output has size (size(img, 1) + padsize[1] + padsize[2], size(img, 2) + padsize[3] + padsize[4]).

SPECTrecon.plus1di! — Method

plus1di!(mat2d, mat1d)

Non-allocating mat2d[i, :] += mat1d

SPECTrecon.plus1dj! — Method

plus1dj!(mat2d, mat1d)

Non-allocating mat2d[:, j] += mat1d

SPECTrecon.plus2di! — Method

plus2di!(mat1d, mat2d, i)

Non-allocating mat1d += mat2d[i,:]

SPECTrecon.plus2dj! — Method

plus2dj!(mat1d, mat2d, j)

Non-allocating mat1d += mat2d[:,j]

SPECTrecon.plus3di! — Method

plus3di!(mat2d, mat3d, i)

Non-allocating mat2d += mat3d[i,:,:]

SPECTrecon.plus3dj! — Method

plus3dj!(mat2d, mat3d, j)

Non-allocating mat2d += mat3d[:,j,:]

SPECTrecon.plus3dk! — Method

plus3dk!(mat2d, mat3d, k)

Non-allocating mat2d += mat3d[:,:,k]

SPECTrecon.project! — Method

project!(views, image, plan, workarray; index)

Project image into multiple views with indexes index (defaults to 1:nview). The 3D views array must be pre-allocated, but need not be initialized.

SPECTrecon.project! — Method

project!(view, plan, workarray, image, viewidx)

SPECT projection of image into a single view with index viewidx. The view must be pre-allocated but need not be initialized to zero.

SPECTrecon.project — Method

views = project(image, mumap, psfs, dy; interpidx, kwargs...)

Convenience method for SPECT forward projector that does all allocation including initializing plan and workarray.

In

image : 3D array [nx,ny,nz]

mumap : [nx,ny,nz] 3D attenuation map, possibly zeros()

psfs : 4D PSF array

dy::RealU : pixel size

Option

interpidx : 1 or 2

SPECTrecon.project — Method

views = project(image, plan, workarray ; kwargs...)

Convenience method for SPECT forward projector that allocates and returns views.

SPECTrecon.rot180! — Method

rot180!(B::AbstractMatrix, A::AbstractMatrix)

In place version of rot180, returning rotation of A in B.

SPECTrecon.rot_f90! — Method

rot_f90!(output, img, m)

In-place version of rotating an image by 90/180/270 degrees

SPECTrecon.rot_f90_adj! — Method

rot_f90_adj!(output, img, m)
The adjoint of rotating an image by 90/180/270 degrees

SPECTrecon.rotate_x! — Method

rotate_x!(output, img, tan_θ, xi, yi, interp, workvec, c_y)

Rotate a 2D image along x axis in clockwise direction using 1d linear interpolation, storing results in output. Sample locations xi and yi must be in increasing order.

SPECTrecon.rotate_x_adj! — Method

rotate_x_adj!(output, img, tan_θ, xi, yi, interp, workvec, c_y)

The adjoint of rotating a 2D image along x axis in clockwise direction using 1d linear interpolation, storing results in output xi and yi must be in increasing order.

SPECTrecon.rotate_y! — Method

rotate_y!(output, img, sin_θ, xi, yi, interp, workvec, c_x)

Rotate a 2D image along y axis in clockwise direction using 1d linear interpolation, storing results in output xi and yi must be in increasing order.

SPECTrecon.rotate_y_adj! — Method

rotate_y_adj!(output, img, sin_θ, xi, yi, interp, workvec, c_x)

Adjoint of rotating a 2D image along y axis in clockwise direction using 1d linear interpolation, storing results in output

SPECTrecon.rotl90! — Method

rotl90!(B::AbstractMatrix, A::AbstractMatrix)

In place version of rotl90, returning rotation of A in B.

SPECTrecon.rotr90! — Method

rotr90!(B::AbstractMatrix, A::AbstractMatrix)

In place version of rotr90, returning rotation of A in B.

SPECTrecon.scale3dj! — Method

scale3dj!(mat2d, mat3d, j, s)

Non-allocating mat2d = s * mat3d[:,j,:]